requirements.txt
anthropic>=0.18.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
resend>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
